+++ app/src/main/java/nextstep/signup/ui/component/SignUpTextField.kt
@@ -5,11 +5,14 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.LocalTextStyle
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.VisualTransformation
+import nextstep.signup.ui.ThemePreviews
+import nextstep.signup.ui.theme.SignupTheme
 
 @Composable
 fun SignUpTextField(
@@ -38,6 +41,31 @@ fun SignUpTextField(
                     )
                 }
             }
+        )
+    }
+}
+
+@ThemePreviews
+@Composable
+private fun SignUpTextFieldPreview() {
+    SignupTheme {
+        SignUpTextField(
+            value = "이지훈",
+            onValueChange = {},
+            label = { Text(text = "이름") }
+        )
+    }
+}
+
+@ThemePreviews
+@Composable
+private fun SignUpTextFieldErrorPreview() {
+    SignupTheme {
+        SignUpTextField(
+            value = "이지훈",
+            onValueChange = {},
+            label = { Text(text = "이름") },
+            supportingText = { Text(text = "error") }
         )
     }
 }
